+++ website-dev/app/routes/home.tsx
@@ -1,5 +1,3 @@
-import { RouteShellPage } from "@/app/components/pages/route-shell-page";
-
 export default function HomePage() {
-  return <RouteShellPage title="Home" />;
+  return null;
 }
