--- website-dev/app/routes/railyard.tsx
@@ -1,5 +1,0 @@
-import { RouteShellPage } from "@/app/components/pages/route-shell-page";
-
-export default function RailyardPage() {
-  return <RouteShellPage title="Railyard" />;
-}
