--- website-dev/app/routes/registry.tsx
@@ -1,5 +1,0 @@
-import { RouteShellPage } from "@/app/components/pages/route-shell-page";
-
-export default function RegistryPage() {
-  return <RouteShellPage title="Registry" />;
-}
