--- website-dev/app/routes/template-mod.tsx
@@ -1,5 +1,0 @@
-import { RouteShellPage } from "@/app/components/pages/route-shell-page";
-
-export default function TemplateModPage() {
-  return <RouteShellPage title="Template Mod" />;
-}
