--- website-dev/app/routes/website.tsx
@@ -1,5 +1,0 @@
-import { RouteShellPage } from "@/app/components/pages/route-shell-page";
-
-export default function WebsiteSuitePage() {
-  return <RouteShellPage title="Website" />;
-}
